api/requirements.txt
fastapi
mangum
aws-lambda-powertools
pydantic[email]
boto3
requests
python-dotenv
passlib[bcrypt]
PyJWT
bcrypt==4.3.0
pytest
httpx 
pytest-cov